+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Version 1.4.0 (destructive)
 
 - changed the audio capture interface
+- added an audio capture service for reading from InputStream
 - close check on the setOnCaptureAudioListener
 
 ## Version 1.3.1